refactor(core): Remove angular from cluster filter service
