+++ app/scripts/modules/core/src/cluster/index.ts
@@ -2,4 +2,4 @@ export * from './ClusterPodTitleWrapper';
 export * from './ClusterRuleMatcher';
 export * from './cluster.service';
 export * from './task.matcher';
-export * from './filter/clusterFilter.service';
+export * from './filter/ClusterFilterService';
