+++ app/scripts/modules/core/src/pipeline/config/stages/deploy/deployExecutionDetails.controller.js
@@ -3,19 +3,18 @@
 import moment from 'moment';
 import _ from 'lodash';
 
-import { CLUSTER_FILTER_SERVICE } from 'core/cluster/filter/clusterFilter.service';
 import { CLOUD_PROVIDER_REGISTRY } from 'core/cloudProvider/cloudProvider.registry';
 import { NameUtils } from 'core/naming';
 import { EXECUTION_DETAILS_SECTION_SERVICE } from 'core/pipeline/details/executionDetailsSection.service';
 import { SERVER_GROUP_READER } from 'core/serverGroup/serverGroupReader.service';
 import { URL_BUILDER_SERVICE } from 'core/navigation/urlBuilder.service';
+import { ClusterState } from 'core/state';
 
 let angular = require('angular');
 
 module.exports = angular
   .module('spinnaker.core.pipeline.stage.deploy.details.controller', [
     require('@uirouter/angularjs').default,
-    CLUSTER_FILTER_SERVICE,
     EXECUTION_DETAILS_SECTION_SERVICE,
     URL_BUILDER_SERVICE,
     CLOUD_PROVIDER_REGISTRY,
@@ -26,7 +25,6 @@ module.exports = angular
     $stateParams,
     executionDetailsSectionService,
     urlBuilderService,
-    clusterFilterService,
     cloudProviderRegistry,
     serverGroupReader,
   ) {
@@ -165,7 +163,7 @@ module.exports = angular
       }
     }
 
-    this.overrideFiltersForUrl = r => clusterFilterService.overrideFiltersForUrl(r);
+    this.overrideFiltersForUrl = r => ClusterState.filterService.overrideFiltersForUrl(r);
 
     let initialize = () => executionDetailsSectionService.synchronizeSection($scope.configSections, initialized);
